move more utilities to legacy
