+++ src/example-types/legacy/terms_stats.js
@@ -1,5 +1,5 @@
 var _ = require('lodash/fp')
-var esTwoLevel = require('../esTwoLevelAggregation').result
+var esTwoLevel = require('./esTwoLevelAggregation').result
 let { buildRegexQueryForWords } = require('../../regex')
 let { getField } = require('../../fields')
